BlockChain check block does not already exists
The check was only enabled for the fully validating blockchain for
performance issues. In our case we keep the blockchain in a contract’s
memory, so it is cheaper to check for the block existence here than to
continue execution and find out the block already exists later on. This
change was motivated by a already-fixed bug that used to make
federators send to the bridge 500 very old bitcoin blocks the bridge
already knows about and trying to add those blocks was very slow