Create FlyoverRedeemScriptParser Class
For every different federation type (standard multisig, legacy erp, p2sh) there is a specific flyover redeem script parser. This is unnecessary, since a flyover redeem script does not care about the federation redeem script format. It can take any redeem script, and it only needs the derivation hash and OP_DROP op code on top.

Create a FlyoverRedeemScriptParser class that implements the RedeemScriptParser interface and has a RedeemScriptParser attribute. It's the RedeemScriptParserFactory responsibility to identify the script as a flyover redeem script, remove the first 2 chunks from it and parse the remaining chunks. FlyoverRedeemScriptParser constructor will receive the full list of chunks that form the redeem script along with a RedeemScriptParser instance that represents the internal redeem script.