Remove createP2shErpRedeemScript  method from P2shErpFederationRedeemScriptParser
At this time we need to remove methods that RedeemScriptUtils already covered so we will remove and refactor tests and look at RSKj to see the impact this has on it.

- Remove createP2shErpRedeemScript  method from P2shErpFederationRedeemScriptParser

- Remove tests related to createP2shErpRedeemScript.

- Ensure all tests are running.

Important Note: All the tests removed will be analyzed and moved to a new class called RedeemScriptUtilsTest.java, to cover the createP2shErpRedeemScript method with tests.